Some documentation, switch assembly default back to cellwise
